requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.6.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.2.0
